Remove i32 and u32 variants of DataType
Curretly noria uses four integer types, i32, u32, i64 and u64 internally,
this is in contrast to SQL that also has i/u8 and i/u16 bit variants, so
it is not entirely clear what i32 and u32 bring to the table on 64 bit
machines.

Performance: on 64 bit machines all operations are performed on 64bits anyway,
             moreover having fewer types greatly reduces the number of branches
             in our logic which should lead to improved performance.
Correctness: if we can get correct results for i8 and i16 using i32, no reason
             we can't use i64 for i32 (same for unsigned variants).
Serialization: all integer types are serialized as a variable length i128, and
               deserialized to the closest type, which in fact can lead to more
               logic bugs.

Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1836
Tested-by: Buildkite CI

+++ noria-client/src/backend/noria_connector.rs
@@ -1338,8 +1338,8 @@ impl NoriaConnector {
                 .get_noria_view(name, region.as_deref(), view_failed)
                 .await?;
 
-            if (offset.is_some() && !matches!(offset, Some(DataType::BigInt(0))))
-                || matches!(limit, Some(DataType::BigInt(0)))
+            if (offset.is_some() && !matches!(offset, Some(DataType::Int(0))))
+                || matches!(limit, Some(DataType::Int(0)))
             {
                 short_circuit_empty_resultset(getter).await
             } else {
